Add files via upload
IFA is a football based game developed by game giants EA Sports every year. So, it was no surprise that EA Sports launched FIFA 19 earlier this year. FIFA 19 contains a rich dataset with plenty of attributes covering all aspects of a real-life footballer in an attempt to immitate him as much as possible in the virtual world. This rich dataset provides a huge oppurtunity for us, data scientists or data analysts (you choose the term !) to analyze and come up with visualizations and patterns